Fix added integer int

+++ avs_core/core/interface.cpp
@@ -616,10 +616,10 @@ int VideoFrame::CheckMemory() const {
   if (vfb->data && vfb->device->device_type == DEV_TYPE_CPU) {
     // check buffer overrun
     int *pInt = (int *)(vfb->data + vfb->data_size);
-    if (pInt[0] != 0xDEADBEEF ||
-      pInt[1] != 0xDEADBEEF ||
-      pInt[2] != 0xDEADBEEF ||
-      pInt[3] != 0xDEADBEEF)
+    if (pInt[0] != (int)0xDEADBEEF ||
+      pInt[1] != (int)0xDEADBEEF ||
+      pInt[2] != (int)0xDEADBEEF ||
+      pInt[3] != (int)0xDEADBEEF)
     {
       return 1;
     }
